+++ src/core/public/utils/index.ts
@@ -31,5 +31,9 @@
 export { shareWeakReplay } from './share_weak_replay';
 export { Sha256 } from './crypto';
 export { MountWrapper, mountReactNode } from './mount';
-export { getWorkspaceIdFromUrl, formatUrlWithWorkspaceId } from './workspace';
-export { WORKSPACE_PATH_PREFIX, WORKSPACE_TYPE } from '../../utils';
+export {
+  WORKSPACE_PATH_PREFIX,
+  WORKSPACE_TYPE,
+  formatUrlWithWorkspaceId,
+  getWorkspaceIdFromUrl,
+} from '../../utils';
