+++ src/core/utils/index.ts
@@ -38,3 +38,4 @@ export {
 } from './context';
 export { DEFAULT_APP_CATEGORIES } from './default_app_categories';
 export { WORKSPACE_PATH_PREFIX, WORKSPACE_TYPE } from './constants';
+export { getWorkspaceIdFromUrl, formatUrlWithWorkspaceId, cleanWorkspaceId } from './workspace';
